device list ui part 3.1
