--- device/src/main/java/com/example/device/data/WillRemowe.kt
@@ -1,3 +1,0 @@
-package com.example.device.data
-
-class WillRemowe
